requirements.txt
attrs==18.2.0
Babel==2.6.0
certifi==2018.10.15
chardet==3.0.4
cliff==2.14.0
cmd2==0.9.6
colorama==0.4.0
idna==2.7
pbr==5.1.1
prettytable==0.7.2
pyparsing==2.3.0
pyperclip==1.7.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
ruamel.yaml==0.15.78
six==1.11.0
stevedore==1.30.0
tosca-parser==1.2.0
urllib3==1.24.1
wcwidth==0.1.7
